Add ability to listen for triggers.
Remove unnecessary tasks. Now it's sorted by name and then latest task is executed.

TODO: Need to check if it's active.

WIP: New way of fixing TaskServiceImpl with update triggers.

WIP: New way of fixing TaskServiceImpl with update triggers.

Add ability to handle triggers.

Triggers are a way to trigger a task based on an event.